lint(test): add helper for child process with output
